requirements.txt
# Core dependencies
numpy>=1.24.0,<2.0.0
pyyaml>=6.0
tqdm>=4.65.0
colorama>=0.4.6

# Optional (for debugging)
matplotlib>=3.7.0  # Visualization of probability distributions
pandas>=2.0.0      # Data analysis utilities

# Development dependencies (optional)
pytest>=7.3.0      # Testing framework
black>=23.0.0      # Code formatting
